API Logout, authentication middleware

+++ src/services/cardService.js
@@ -5,45 +5,57 @@ import { columnModel } from '~/models/columnModel'
 import ApiError from '~/utils/ApiError'
 import { StatusCodes } from 'http-status-codes'
 
-const createNew = async (reqBody) => {
+const createNew = async (reqBody, userInfo) => {
   try {
     const data = {
-      ...reqBody
+      ...reqBody,
+      owner: userInfo.username
     }
-
+    // const board = await boardModel.findOneById(data.boardId)
+    // if (userInfo.username !== board.owner) {
+    //   throw new ApiError(StatusCodes.FORBIDDEN, 'User is unauthorized')
+    // }
     // Gọi tới tầng model để xử lý lưu bản ghi newcolumn vào DB
-    const createdCard = await cardModel.createNew(data)
-    // console.log(createdcolumn)
+    const result = await cardModel.createNew(data)
 
     // Trả về bản ghi mới cho phía client
-    const result = await cardModel.findOneById(createdCard.insertedId)
+    const createdCard = await cardModel.findOneById(result.insertedId)
 
     //xử lí dữ liệu cho chuẩn với FE
-    if (result) {
-      columnModel.pushCardOrderIds(result)
-      columnModel.update(result.columnId, { updatedAt: Date.now() })
-      boardModel.update(result.boardId, { updatedAt: Date.now() })
+    if (createdCard) {
+      columnModel.pushCardOrderIds(createdCard)
+      columnModel.update(createdCard.columnId, { updatedAt: Date.now() })
+      boardModel.update(createdCard.boardId, { updatedAt: Date.now() })
     }
-    return result
+    return createdCard
   } catch (error) { throw error }
 }
 
-const deleteCard = async (cardId) => {
+const deleteCard = async (cardId, userInfo) => {
   try {
     const cardToDelete = await cardModel.findOneById(cardId)
     if (!cardToDelete) {
       throw new ApiError(StatusCodes.NOT_FOUND, 'Card not found')
     }
+
+    // const board = await boardModel.findOneById(cardToDelete.boardId)
+    // if (userInfo.username !== board.owner) {
+    //   throw new ApiError(StatusCodes.FORBIDDEN, 'User is unauthorized')
+    // }
+    if (userInfo.username !== cardToDelete.owner) {
+      throw new ApiError(StatusCodes.FORBIDDEN, 'User is unauthorized')
+    }
+
     // Xóa card
-    await cardModel.deleteOneById(cardId)
+    cardModel.deleteOneById(cardId)
     // Update cardOrderIds array for column
-    await columnModel.pullCardOrderIds(cardToDelete)
+    columnModel.pullCardOrderIds(cardToDelete)
 
     // Update updatedAt field for column and board
-    await columnModel.update(cardToDelete.columnId, {
+    columnModel.update(cardToDelete.columnId, {
       updatedAt: Date.now()
     })
-    await boardModel.update(cardToDelete.boardId, {
+    boardModel.update(cardToDelete.boardId, {
       updatedAt: Date.now()
     })
 
@@ -51,19 +63,32 @@ const deleteCard = async (cardId) => {
   } catch (error) { throw error }
 }
 
-const update = async (cardId, reqBody) => {
+const update = async (cardId, reqBody, userInfo) => {
   try {
+    const cardToUpdate = await cardModel.findOneById(cardId)
+    if (!cardToUpdate) {
+      throw new ApiError(StatusCodes.NOT_FOUND, 'Card not found')
+    }
+
+    // const board = await boardModel.findOneById(cardToUpdate.boardId)
+    // if (userInfo.username !== board.owner) {
+    //   throw new ApiError(StatusCodes.FORBIDDEN, 'User is unauthorized')
+    // }
+    if (userInfo.username !== cardToUpdate.owner) {
+      throw new ApiError(StatusCodes.FORBIDDEN, 'User is unauthorized')
+    }
+
     const dataToUpdate = {
       ...reqBody,
       updatedAt: Date.now()
     }
 
-    const result = await cardModel.update(cardId, dataToUpdate)
-    if (result) {
-      columnModel.update(result.columnId, { updatedAt: Date.now() })
-      boardModel.update(result.boardId, { updatedAt: Date.now() })
+    const updatedCard = await cardModel.update(cardId, dataToUpdate)
+    if (updatedCard) {
+      columnModel.update(updatedCard.columnId, { updatedAt: Date.now() })
+      boardModel.update(updatedCard.boardId, { updatedAt: Date.now() })
     }
-    return result
+    return updatedCard
   } catch (error) { throw error }
 }
